Version 1.9 of PlayVideo
If you prepend the file name with the @ sign in the list.lst file, it will make that video loop indefinitely (until you advance to a new video).

Fixed bug in PlayVideo that caused files greater than 2.147 GB to not be found. (fopen() replaced with fopen64())

ListManager now tries 6 times to open the list directory